requirements.txt
streamlit>=1.40.0
pandas>=2.2.0
networkx>=3.2.0
plotly>=5.24.0
fpdf2>=2.7.0
neo4j>=5.24.0
python-dotenv>=1.0.1
requests>=2.32.0
feedparser>=6.0.11
beautifulsoup4>=4.12.3
